requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
dnspython==2.2.0
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.73.0
fastapi-mail==1.0.4
h11==0.13.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-metadata==4.11.1
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pydantic==1.9.0
pyparsing==3.0.7
python-decouple==3.6
python-multipart==0.0.5
redis==4.1.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.5
wrapt==1.13.3
zipp==3.7.0